+++ lib/presentation/preview_screen.dart
@@ -3,7 +3,6 @@ import 'package:dict/presentation/hamburger_menu.dart';
 import 'package:dict/presentation/widgets/beat_widget.dart';
 import 'package:dict/presentation/widgets/carousel_widget.dart';
 import 'package:dict/presentation/widgets/my_app_bar.dart';
-import 'package:dict/presentation/widgets/staff.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -20,42 +19,37 @@ class PreviewScreen extends StatelessWidget {
       body: Column(
         children: [
           CarouselWidget(width: width, mainScale: mainScale),
-          SizedBox(
-            height: mainScale * 24 + 12 * mainScale,
-            child: Stack(
-              children: [
-                // TODO(zoskar): change order
-                BlocBuilder<DictandoCubit, DictandoState>(
-                  builder: (context, state) => BeatWidget(
-                    step: mainScale,
-                    preview: true,
-                  ),
-                ),
-                Staff(distance: mainScale, isTappable: false),
-              ],
-            ),
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: [
-              FloatingActionButton(
-                onPressed: () {
-                  context.read<DictandoCubit>().beatLeft();
-                },
-                heroTag: null,
-                child: const Icon(Icons.arrow_back),
-              ),
-              FloatingActionButton(
-                onPressed: () {
-                  context.read<DictandoCubit>().beatRight(preview: true);
-                },
-                heroTag: null,
-                child: const Icon(Icons.arrow_forward),
-              )
-            ],
-          )
+          BeatWidget(mainScale: mainScale),
+          const BeatChangeArrows()
         ],
       ),
     );
   }
+}
+
+class BeatChangeArrows extends StatelessWidget {
+  const BeatChangeArrows({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) => Row(
+        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+        children: [
+          FloatingActionButton(
+            onPressed: () {
+              context.read<DictandoCubit>().beatLeft();
+            },
+            heroTag: null,
+            child: const Icon(Icons.arrow_back),
+          ),
+          FloatingActionButton(
+            onPressed: () {
+              context.read<DictandoCubit>().beatRight(preview: true);
+            },
+            heroTag: null,
+            child: const Icon(Icons.arrow_forward),
+          )
+        ],
+      );
 }
